TextBox, List abstraction
