+++ src/main/emulator/lcdui/IListImpl.java
@@ -1,0 +1,9 @@
+package emulator.lcdui;
+
+import javax.microedition.lcdui.Choice;
+import javax.microedition.lcdui.Graphics;
+
+public interface IListImpl extends Choice, IScreenImpl {
+
+	void drawScrollBar(Graphics graphics);
+}
